Refactor package structure to org.jetbrains.kotlinx.kandy.contexts
